Refactorings and cleanups
