+++ src/main/java/com/teamwizardry/wizardry/common/module/events/ModuleEventCollideEntity.java
@@ -2,9 +2,10 @@ package com.teamwizardry.wizardry.common.module.events;
 
 import com.teamwizardry.wizardry.api.spell.SpellData;
 import com.teamwizardry.wizardry.api.spell.SpellRing;
+import com.teamwizardry.wizardry.api.spell.annotation.RegisterModule;
 import com.teamwizardry.wizardry.api.spell.module.IModuleEvent;
 import com.teamwizardry.wizardry.api.spell.module.ModuleEvent;
-import com.teamwizardry.wizardry.api.spell.module.RegisterModule;
+
 import net.minecraft.entity.Entity;
 import javax.annotation.Nonnull;
