+++ src/main/java/com/teamwizardry/wizardry/common/module/modifiers/ModuleModifierIncreaseAOE.java
@@ -1,9 +1,7 @@
 package com.teamwizardry.wizardry.common.module.modifiers;
 
+import com.teamwizardry.wizardry.api.spell.annotation.RegisterModule;
 import com.teamwizardry.wizardry.api.spell.module.IModuleModifier;
-import com.teamwizardry.wizardry.api.spell.module.RegisterModule;
-
-import javax.annotation.Nonnull;
 
 /**
  * Created by Demoniaque.
